# Optimize SSE CRC