@MichalPetryka 2020-09-28
I've also tested a modified version of the managed code, it seemed to be about 5% slower on .Net Core 3.1 but about 5% faster on .Net 5.0, don't think it's worth changing though.